// backend/src/modules/auth/auth.schemas.ts
import { z } from 'zod';

export const registerSchema = z.object({
  name: z.string().min(2).max(80),
  email: z.string().email(),
  password: z.string().min(8),
  role: z.enum(['USER', 'SELLER']).default('USER'),
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(1),
});